refactor: Code cleanup and minor improvements
CatalogRepository:
- Seal class for devirtualization
- Use Environment.CurrentManagedThreadId instead of Thread.CurrentThread
- Remove unused using directive
- Use collection expressions for cleaner syntax
- Make Dispose method private (only called through IDisposable pattern)

CreateCacheCommand:
- Convert to record type for immutability

StringExtension:
- Remove unused IsNullOrEmpty extension method
- Refactor MakeRelativePath with improved readability
- Enhance code comments for clarity
- Move to cdeLib.Extensions namespace for consistency
