+++ src/cdeLib/Extensions/StringExtension.cs
@@ -1,16 +1,11 @@
 ﻿using System;
 
-namespace cdeLib;
+namespace cdeLib.Extensions;
 
 public static class StringExtension
 {
-    public static bool IsNullOrEmpty(this string value)
-    {
-        return String.IsNullOrEmpty(value);
-    }
-
     /// <summary>
-    /// Assumption that rootPath of form "D:" is equivalent to root of device not a relative path to start with ?
+    /// Assumption that rootPath of form "D:" is equivalent to root of a device not a relative path to start with?
     /// </summary>
     /// <param name="fullPath"></param>
     /// <param name="rootPath"></param>
@@ -28,14 +23,14 @@ public static class StringExtension
         }
 
         // Use Span for zero-allocation path manipulation
-        ReadOnlySpan<char> fullPathSpan = fullPath.AsSpan();
-        ReadOnlySpan<char> rootPathSpan = rootPath.AsSpan();
+        var fullPathSpan = fullPath.AsSpan();
+        var rootPathSpan = rootPath.AsSpan();
 
-        // Check if we need to add directory separator
-        int rootLength = rootPath.Length;
+        // Check if we need to add a directory separator
+        var rootLength = rootPath.Length;
         if (!System.IO.Path.EndsInDirectorySeparator(rootPath))
         {
-            // Check if fullPath starts with rootPath + separator
+            // Check if the fullPath starts with rootPath + separator
             if (fullPathSpan.Length > rootLength &&
                 fullPathSpan.StartsWith(rootPathSpan, StringComparison.Ordinal) &&
                 (fullPathSpan[rootLength] == '\\' || fullPathSpan[rootLength] == '/'))
@@ -43,15 +38,13 @@ public static class StringExtension
                 // Skip rootPath and the separator
                 return new string(fullPathSpan[(rootLength + 1)..]);
             }
+
             return null;
         }
 
         // rootPath already ends with separator
-        if (fullPathSpan.StartsWith(rootPathSpan, StringComparison.Ordinal))
-        {
-            return new string(fullPathSpan[rootLength..]);
-        }
-
-        return null;
+        return fullPathSpan.StartsWith(rootPathSpan, StringComparison.Ordinal)
+            ? new string(fullPathSpan[rootLength..])
+            : null;
     }
 }
